Reduce branches differing for GLES3 vs. GL 3.3
I want to see if I can make a cursed GLAD header that works on both
OpenGL ES 3.0 and OpenGL 3.3.
GLES 3 has less functionality and is based on GL 3.3 Core. But, there
are a bunch of functions and definitions shared between the two.
Some definitions or functions would need to be stubbed or redefined for
GLES3, but since this thing already works on GLES3 it shouldn't even be
too difficult to go from there.
The hard part may just be making that one cursed GLAD header.